[feat] #35 - Move repository directory
