+++ src/modules/auth/auth.module.ts
@@ -5,8 +5,8 @@ import { ConfigModule, ConfigService } from '@nestjs/config';
 import { PassportModule } from '@nestjs/passport';
 import { JwtModule } from '@nestjs/jwt';
 import { TypeormRepositoryModule } from '../../database/typeorm-repository.module';
-import { UserRepository } from '../../database/repositories/user.repository';
-import { AdminRepository } from '../../database/repositories/admin.repository';
+import { UserRepository } from '../../repositories/user.repository';
+import { AdminRepository } from '../../repositories/admin.repository';
 import { SmtpConfig } from '../../config/smtp.config';
 import { JwtStrategy } from './jwt/jwt.strategy';
 import { RefreshStrategy } from './jwt/refresh.strategy';
