+++ src/modules/auth/jwt/jwt.strategy.ts
@@ -2,7 +2,7 @@ import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { ExtractJwt, Strategy } from 'passport-jwt';
 import { ConfigService } from '@nestjs/config';
-import { UserRepository } from '../../../database/repositories/user.repository';
+import { UserRepository } from '../../../repositories/user.repository';
 import { JwtPayload } from './jwt.payload';
 
 @Injectable()
