+++ src/modules/searches/searches.module.ts
@@ -2,10 +2,10 @@ import { Module } from '@nestjs/common';
 import { SearchesController } from './searches.controller';
 import { SearchesService } from './searches.service';
 import { TypeormRepositoryModule } from '../../database/typeorm-repository.module';
-import { HashTagRepository } from '../../database/repositories/hash-tag.repository';
-import { EstateRepository } from '../../database/repositories/estate.repository';
-import { ProposalRepository } from '../../database/repositories/proposal.repository';
-import { LinkingRepository } from '../../database/repositories/linking.repository';
+import { HashTagRepository } from '../../repositories/hash-tag.repository';
+import { EstateRepository } from '../../repositories/estate.repository';
+import { ProposalRepository } from '../../repositories/proposal.repository';
+import { LinkingRepository } from '../../repositories/linking.repository';
 
 @Module({
   imports: [
