+++ src/modules/users/users.module.ts
@@ -2,21 +2,21 @@ import { Module } from '@nestjs/common';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
 import { TypeormRepositoryModule } from '../../database/typeorm-repository.module';
-import { UserRepository } from '../../database/repositories/user.repository';
+import { UserRepository } from '../../repositories/user.repository';
 import {
   EstateLikeRepository,
   EstateRepository,
-} from '../../database/repositories/estate.repository';
+} from '../../repositories/estate.repository';
 import {
   OpinionRepository,
   ProposalLikeRepository,
   ProposalRepository,
-} from '../../database/repositories/proposal.repository';
+} from '../../repositories/proposal.repository';
 import {
   LinkingLikeRepository,
   LinkingRepository,
   LinkingRequestRepository,
-} from '../../database/repositories/linking.repository';
+} from '../../repositories/linking.repository';
 
 @Module({
   imports: [
