+++ src/modules/users/users.service.ts
@@ -6,7 +6,7 @@ import {
   UnauthorizedException,
 } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
-import { UserRepository } from '../../database/repositories/user.repository';
+import { UserRepository } from '../../repositories/user.repository';
 import { adjectives, nouns } from './nickname-keywords';
 import { UpdateUserinfoReqDto } from './dto/req/update-userinfo-req.dto';
 import * as bcrypt from 'bcrypt';
