+++ src/repositories/hash-tag.repository.ts
@@ -1,6 +1,6 @@
-import { TypeormRepository } from '../typeorm-repository.decorator';
+import { TypeormRepository } from '../database/typeorm-repository.decorator';
 import { Repository } from 'typeorm';
-import { HashTagEntity } from '../entities';
+import { HashTagEntity } from '../database/entities';
 
 @TypeormRepository(HashTagEntity)
 export class HashTagRepository extends Repository<HashTagEntity> {}
